+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - `process_file` & `process_all_file` — Added New parameter `to_list` to be real python list if you don't need array
 - Added `auto_fix` function for automatic data type correction in DataFrames for mixed Dtypes.
 - Added logg to `auto_fix` to track changes made to columns.
+- Added Documentation for `auto_fix` function to understand more about function.
 
 ---